Se sobrecargo el metodo formatTime de la clase TranslatorHelper para poder pasarle como parametro un long, ya que aveces con el entero, causa desbordamiento.

+++ Reportes2ASPEngine/MiddleWareReports/reports/GenericReport.cs
@@ -421,7 +421,7 @@ namespace MiddleWareReports
                         string value = TranslatorHelper.parseDbValue(dataRow[column.ColumnName]);
                         if (convertedColumns[column.ColumnName] != null && value != "")
                         {
-                            value = TranslatorHelper.formatTime(Convert.ToInt32(value));
+                            value = TranslatorHelper.formatTime(Convert.ToInt64(value));
                         }
                         XmlElement el = xmlReport.CreateElement("", "Cell", "");
